fix(MCOL-4611): MOD loses precision on huge narrow decimal

+++ utils/funcexp/func_mod.cpp
@@ -106,7 +106,7 @@ IDB_Decimal Func_mod::getDecimalVal(Row& row, FunctionParm& parm, bool& isNull,
   }
 
   IDB_Decimal d = parm[0]->data()->getDecimalVal(row, isNull);
-  int64_t value = d.value / pow(10.0, d.scale);
+  int64_t value = d.value / static_cast<int64_t>(pow(10.0, d.scale));
   int lefto = d.value % (int)pow(10.0, d.scale);
 
   int64_t mod = (value % div) * pow(10.0, d.scale) + lefto;
